1) In GET4 analysis treat correctly rising and falling edges. 2) Make correct calculations of signal width in get4test 3) Add leading edge difference to next channel in get4test 4) Add possibility to ignore 250MHz times in get4 processor    to avoid problems with unsynchronized clocks
git-svn-id: https://subversion.gsi.de/go4/app/stream@928 302e7798-3865-0410-85c9-d38765bbeff6

+++ framework/base/StreamProc.cxx
@@ -464,7 +464,7 @@ bool base::StreamProc::SkipBuffers(unsigned num_skip)
    // local triggers must be cleanup by other means,
    // TODO: if it would not be possible, one could use front time of buffers to skip triggers
    if (fLocalTrig.size()>1000)
-      printf("Too much %u local trigger remaining - why\n?", fLocalTrig.size());
+      printf("Too much %u local trigger remaining - why\n?", (unsigned) fLocalTrig.size());
 
    if (fQueueScanIndex>=num_skip) {
       fQueueScanIndex-=num_skip;
